Kill `ZipKernel for VarBinViewVTable`
