+++ vortex-array/src/arrays/varbinview/compute/mod.rs
@@ -8,8 +8,6 @@ mod is_sorted;
 mod mask;
 mod min_max;
 mod take;
-mod zip;
-
 #[cfg(test)]
 mod tests {
     use vortex_buffer::buffer;
